Add trailing slash to share URL (for counting purposes only)
Twitter normalizes the URL when getting the count with a trailing slash, which is great. Unfortunately, they return this normalized URL instead of the URL you asked for. This causes inaccurate share counts for Twitter if you don't have trailing slashes. Adding a trailing slash to the URL in the `WPCOM_sharing_counts` array fixes the issue.

I verified that this doesn't break Facebook share counts. I couldn't test with LinkedIn.