Protect Module: Improving IP detection
In BrutePrtoect, the `get_ip` function hinged largely on the existence
of php's `filter_var` function. since Jetpack supports back to php 5.2,
and `filter_var` existed in 5.2, we can loose that check altogether.
Additionally, the `is_ip_private` was only supporting ipv4, and i've
fixed it to support ipv6 as well.