REST API: fix the acceptance of a value of `false` in the publicize update connection endpoints
The casting function in the REST API does not interpret `(string|bool)` properly and as such passing in `false` as the value of `external_user_ID` was not working correctly. I'm introducing a `WPCOM_JSON_API::is_falsy` method which is a counterpart to `WPCOM_JSON_API::is_truthy` to determine if a passed in string represents a false value. The update publicize endpoints now use that new method to determine if a string of false was passed in. Finally I've updated the documentation for those endpoints to indicate that only a boolean value of `false` is accepted, and not `true`.

Fixes https://github.com/Automattic/io/issues/203

Merges r111923-wpcom.