iCalendar Reader: Resolve inconsistent use of unixtime and $current
Adds the filter `ical_get_current_time` to the initial call to `time()` to allow unit tests to run from a set unixtime.

Updates `date()` and `time()` functions running with the default timestamp `time()` to include this value as the second parameter, ensuring that the entirety of `filter_past_and_recurring_events()` runs from the same value for `$current`.

Merges r112068-wpcom.