REST API Post listing endpoints: Some status parameter fixes for logged-out users
For logged-out users, limit the status parameter to 'publish' or 'any' (the latter being
synonymous with 'publish'), and short-circuit with a quick return if protected statuses
were requested.

This addresses another bug that was causing the "found" parameter to be incorrect.

Post listing v1 and v1.1 endpoints.

https://[private link]

Merges r112151-wpcom.