Fixes repeating posts in case the infinite scroll is activated on a search page
As the search takes post_title match into consideration while ordering, the posts which actually match the search due to a keyword in the title are placed before the rest of the search results on the initial page load and thus they are being repeated during the initial infinite scrolling request as they match even the restricted query with post_date.

See https://[private link] for more details

In order to fix this issue, I added a query_var post__not_in which is being filled in by posts which match the search by title and this one is being passed around during requests. It required some modifications to the logic, so I had to introduce a method get_query_vars which wraps the wp_query()->query_vars and adds mentioned param in case it is appropriate.

Merges r112224-wpcom.