Protect Module: api key visiibility
If the API key is a-ok, there is no reason to expose it
to the user at all. We'll only show API key related dialogs
if there is an error. Errors obtaining an API key will very
likely be related to Jetpack connection errors, so we'll also
include a link to the debug page.