Jetpack Manage: Sync Constants
Updated to separate constants from options.
They should be treated a bit differently then options.
This also makes them a bit easier to work with. Since now to sync a
constant we just add it to the list of constants that need to be synced
or add Jetpack_Sync::sync_constant( __FILE__, 'HELLO_WORLD' ); to the
module that relies on the constant.