REST API: Check capabilities for post types in the /posts endpoint, enable "feedback" CPT
We tell people to filter "rest_api_allowed_post_types" to enable a CPT to
be accessed via the API, but we aren't actually doing any checks to make
sure the requesting user can access the post type in question. Instead of
placing the burden on the user to wrap the filter in a capabilities check,
let's just do the right thing when the request is made.

This also enables the "feedback" CPT (contact form submissions) for API access.

All bundled together because the new test needs something to test.

https://[private link]

Merges r114224-wpcom.