REST API
Previously, we would rely on the is_jetpack_site filter (while executing code on WordPress.com) to see if a certain blog is a Jetpack blog.
We only expected this code to execute on WordPress.com run code (aka /me/posts) rather then by the target Jetpack blog.
However, now Jetpack uses a filter with the same name in some of the syncing code, so the author and featured image code using the filter are incorrect.

This change updates the check to only execute the block if the code is also being executed by .com (and not by the jp blog itsself). That way /me/posts and others are happy, and the Jetpack blog returns the correct info for /sites/$site/posts/* as well.
The bug that this caused was all Jetpack posts (on the jptest site) were returning an author of 0. Since the wrong author code in get_author was running.

Merges r114326-wpcom.