Port r113525-wpcom downstream
Sharing Buttons

On April 30th, Facebook will deprecate old unversioned endpoints. We
can't upgrade this call to v2 of the graph api because those endpoints
are authenticated only.

We are currently investigating some alternatives, but in the meantime
we want the current code (especially in Jetpack) to gracefully degrade
when the date rolls around.

Props @justinshreve