Testimonial Plugin: correctly set featured image setting in Customizer
Add a filter for image ID -> GUID so that when the featured image attachment is
sent to the javascript, it sends the URL and not the GUID, as this tool operates
on the URL (the inverse function, to go from GUID -> ID already exists, but is
improved by this commit also -- it now searches all posts for the attachment
guid and not just those within the last 10 posts returned by get_posts).

Part of https://[private link]

Still does not fix the bug that prevents the image from showing in the control
on load, but hopefully gets us closer.

Merges r114772-wpcom.