iCalendar Reader: Fix repeating month and year events.
When an event is set to repeat, we've neglected to prepend the current year, or year and month combination, which shifted the event to create a strange bug with yearly events repeating frequently, but with a time that reflected the date-range of the event.

See https://[private link]

Merges r115751-wpcom.