iCalendar Reader: Catch DateTime conversion errors.
DateTime is used for parsing, storing, and displaying event times. When times are parsed and stored, they are converted to UTC, and when displayed, they are converted to the display timezone.

We need to catch DateTime exceptions when parsing and storing event dates to avoid throwing fatal errors.

Merges r115987-wpcom.