Custom-css: Remove content filters when restoring revisions
When a custom-css change is saved, the content filters are disabled to prevent
mangling the CSS code. However, if a previous version of the CSS is restored,
the restore uses the standard WP post restore process which also includes
passing the content through sanitize_post(). This means that the CSS may not
be restored as intended.

For example, if the CSS contains ampersands (as are used in SCSS), the
ampersands will be transformed into &amp; during the restore process.

This commit removes the content filters for revision restores on the safecss
post_type to resolve that issue.

Fixes #7306

Merges r115994-wpcom.