Signup: Adding a new allow_unauthorized_request property for some endpoints that we want to be accessed by unauthenticated users.
This also includes tests for this change. Since the tests don't make a real http call, I have wrapped the headers setting in a headers_sent check, so that we don't get warnings when running the tests.

This also fixes one of the push notifications tests that was behaving unexpectedly on a sandbox.

See https://[private link] for more details.

Merges r116175-wpcom.