Comments API: Grab post status through `get_post_status()`
Previously there were certain requests to create new comments or
comment replies that failed with the error,

`Comments on drafts are not allowed`

But the parent post or comment wasn't a draft at all. It turns out
that in the API endpoint the code was grabbing the parent post's
status by using `$post->post_status`, which seems reasonable, but
returns the wrong information for pages such as media attachment
pages.

The proper behavior in these cases should be to get the post status of
the parent post from which the media attachment page hangs. The correct
behavior was already present in `wp-comments-post.php` but not in the
API endpoint.

This commit replaces the simpler status call with the helper function
to get the proper value when posting to things like media attachment
pages.

Merges r116196-wpcom.