REST API: current_user_can_access_post_type should check publicly_queryable
In practice, this won't change that much. publicly_queryable defaults to
whatever public is set to. You have to explicitly set publicly_queryable
if you want it to be different.

The use case is if you want a post type to be available in the admin, but
only publicly available where you manually query -- i.e.) not available
in parse_request or the API.

https://[private link]#comment-62633

Merges r116558-wpcom.