Custom Post Types: Carried the settings section definition for Custom Post Types out into a common file.
The section may be used by other custom post types later on when we add them to wpcom, for now
this is done to synchronize the custom-post-types.php file with the Jetpack repository. This is
also the reason for introducing IS_WPCOM checks in portfolio.php - to avoid calling an undefined
stats function in Jetpack.

Props @dereksmart for working on those changes.

Merges r116916-wpcom.