Notes: Stop depending on `jquery.spin` so that `wp-admin/network` loads the notes properly
Issue #718

The notifications panel wasn't loading on `wp-admin/network` pages. This was apparently
due to a conflict between `jquery.spin` and some other library. Removing this library
from the list of dependencies for `notes-common-v2.js` allowed it to load.

Strangely enough, the spinner functionality didn't stop when I removed the dependency.
The library still appears to be loading from another request.

Were this to cause any bad side effects, it could be limited to only impacting the
network admin with a simple conditional `is_network_admin()`