Infinite scroll enhancements: Merge from Jetpack
Stop ignoring media containing elements

Ignoring elements that may contain media sounded nice, but it limited the usefulness of replacing the elements. Audio elements especially tend to cause crashes in Chrome when many are in the page at once, and iframes carry a large memory weight. Tear 'em out.

Props Ben Lowery <blowery@users.noreply.github.com>
Authored on Mon Aug 11 15:42:59 2014 -0400
https://github.com/Automattic/jetpack/commit/f255043

***

Big Whitespace Fix.

Also some minor code style things, adding braces, etc.  But mostly
converting line indentations from spaces to tabs, most of which have
crept in via pull requests.

Props George Stephanis <george@stephanis.info>
Authored on Fri Oct 3 15:04:42 2014 -0400
https://github.com/Automattic/jetpack/commit/e4c9584

***

infinity.js : Resolve Divergence from Jetpack

This patch teaches the infinite scroll system how to pull DOM nodes out of
nodes outside the current viewport and stash them in a cache. It does this
by hooking into determineURL and walking all of the pages that were determined
to be hidden. It pulls all of the child nodes out of the page and stashes them
in a DocumentFragment. It also updates the min-height of the page wrapper to
match the old height when all the children were present.

When a user scrolls a page back into view, the children are reinserted into
the wrapper and the min-height and class marker are removed.

See https://github.com/Automattic/jetpack/pull/300

The cache that was stored did not work with lazy load or any other plugin that fires on the 'post-load' event.
Since the post load wasn't fired every time the content was added or re-added to the dom.
Now we store the html also right before we take it away because it might have been altered by the post-load event.

Another  optimization that was done to fire the change url  in the same timer as the the refresh() since this will only create one times instead of multiple once.

Merges r106807-wpcom.

Props Enej Bajgoric <enej.bajgoric@automattic.com>
Authored on Mon Oct 13 23:34:03 2014 +0000
https://github.com/Automattic/jetpack/commit/c467e74

***

Infinity: added a `button` wrapper for infinity handle

This will allow to focus on the "Older posts" button and trigger it
with the keyboard, for example.

Props Igor Zinovyev <zinigor@gmail.com>
Authored on Sat Dec 13 02:04:22 2014 +0300
https://github.com/Automattic/jetpack/commit/4bf2f11

***

Check for response.html before using indexOf on it.

This is a sanity check and should be there in the off-chance that `response.html` is `null` (yes I have seen it happen). Bottom line, if there is no `response.html`, then it's obvious that media player JS will not be needed.

Props Justin Sternberg <me@jtsternberg.com>
Authored on Thu Dec 18 14:07:52 2014 -0500
https://github.com/Automattic/jetpack/commit/e8158fa

Merges r117721-wpcom.