Add Jetpack_Client:: wpcom_json_api_request_as_blog()
This lets us call specific whitelisted endpoints with authentication
based on the blog connection, rather than a specific user connection.

This is important because it means we can query things like Stats
endpoints and Related Posts endpoints on behalf of the blog, rather
than a specific user.  Things that we want to be authenticated, but
don’t care about the user on.  That way, admins that aren’t connected
to WPCOM will be able to get stats and the like.

👌