Add `stats_get_from_restapi` and caching stuff.
Use the same 5 minute cache used by the old `stats_get_csv()`
functions, but in a new option.

The structure of the return data may change, this is just in place for
playing around at the moment.  It will eventually be used for all in
dashboard stats, as it also pulls down unique visitors and much richer
data, not merely views.