Protect Module: Get local whitelist
Introducing the concept of a local whitelist which will be
stored in a Jetpack Option.

Currently we store the whitelist in a site_option, which on multi-site,
will apply to the entire network. This is not ideal, because in
most cases, we only want to whitelist an IP Address for a single site
in the network, rather than the entire network.

There will now be local whitelists and network whitelists for multi-site
installs. Nothing will change for single site installs which will just
keep their local whitelists.