Show columns setting with TM galleries as default
Previously, the Columns setting was only being shown when the "default" gallery type was being displayed. But when Tiled Mosaic galleries were set as the default under Settings > Media, the "default" type (previously Thumbnail Grid) was renamed to "thumbnails" here:

https://github.com/Automattic/jetpack/blob/3.5.3/modules/tiled-gallery/tiled-gallery.php#L168

So adding a simple check to see if we want a "default" gallery OR a "thumbnails" gallery fixes #1287.