Only check on beta versions
We’re not quite sure how this will behave in all use cases and server
environments — if, for example, hosts records are modified so a public
domain appears to be in a reserved local ip range.

For the moment, only run it on prerelease versions, and not on public
stable releases.