REST API: temporarily comment out relatedposts settings code
We have noticed that we're getting fatal errors for some endpoints in the API because theme functions are loaded twice. We haven't figured it out yet, but in the meantime it's better to just remove these particular settings that rely upon a theme check and allow the settings endpoint to work correctly otherwise.

Merges r120245-wpcom.