unique disconnects not tracked correctly
since we can’t make assumptions about how many times previously
connected sites might have been disconnected we will start with a value
of -1. If the site is later disconnected it skips to a value of 1 so
the disconnect count is accurate.

Unique disconnects will only be tracked if the site was connected for
the first time after 3.6