iCalendar Reader: Increment the COUNT counter for monthly recurring events.
When events start in the distant past with a COUNT parameter, we've been ignoring it until the event is in the future of our current date.

This leads to including long-expired events in our upcoming events calendar; the COUNT limit is never reached until we are in the future, even though the event should have started and incremented COUNT in the first few event checks.

Merges r121577-wpcom.