REST API: Media: Allow empty values to be specified for item fields
Allows title, caption, description, parent_id, alt, artist, and album fields to be set to an empty string.

It's reasonable to expect that a developer may want to remove one or more of these attributes after they've been assigned.

This is not considered to be backwards-incompatible becaues the previous behavior is considered to be a bug.

Merges r121710-wpcom.