NUX: Reset data if page is visited via "back" button
Since we're using ajax to update the elements, if someone navigates away frmo the page and returns via the "back" button, the cached data will not be accurate anymore.

This creates a new ajax function that will re-send the correct data to the templates.  We also manually add a hash to the URL any time an ajax post happens, which is what will trigger the refresh to happen `onload`