Sharing: allow site owners to specify a custom Facebook app ID.
This custom App ID will be used with the official Facebook Share button,
with the Facebook Like Box, and with the Facebook shortcode.

It also avoids having the Facebook Share button display a link to WordPress.com by default.
Instead, the app ID isn't added by default for that button, since it is not required:
https://developers.facebook.com/docs/plugins/share-button