Switch `Jetpack_PostImages::from_gallery()` to using `get_post_galleries_images()`
Removes the custom functionality, switches to using
get_post_galleries_images().

`get_post_galleries_images()` returns a multidimentional array of image
URLs for all the galleries it finds in a post.

Could have used `get_post_galleries()`, decided against it since it
returns gallery specific data which is not used in this case.

The output is exactly the same as before, except now we're using the
built in WP function get the images.