REST API: enable users search via display_name
The /site/$site/users endpoint makes use of WP_User_Query, and the prepare_query() method inside that strips all but a whitelisted set of search_columns. And for some reason display_name is not a whitelisted search column. Changing this would require a core patch. So instead for now I am overriding this with a users_search_columns filter that will apply the requested search_columns to the query (also whitelisted, but includes display_name).

I also noticed that the test case for search_columns was incorrect. It was using an invalid column name, which curiously resulted in the test working because if you *don't* provide a search_columns param (which was the effective case here) and the search term includes an @ sign, user_email is assumed to be the search column. I have corrected the test to specifically apply the user_email search column.

I will pursue a patch in core to allow display_name as a search column for users, and we can hopefully remove this workaround.

More discussion of this change can be found here: https://github.com/Automattic/calypso-pre-oss/pull/9275#issuecomment-134471090

Merges r123202-wpcom.