REST API: In the post update endpoint, when transitioning from a draft to a published post, set the date to now if not supplied in the input.
Otherwise, drafts are published with their last draft date, which can be out of date. Say you've been sitting on a draft for a week and then suddenly publish it using an API call. Before this commit, it would keep its old draft publication date, and any posts published in the mean time would appear to be more recently published.

Merges r123566-wpcom.