Shortcodes: update slideshow gallery shortcode
jQuery.browser was deprecated in jQuery 1.9. We've continued to be able to use it via the jQuery migrate plugin, but this is not a viable option longterm. Additionally this broke instances where we need to load the script without loading the migrate plugin, as was the case in the Calypso Editor.

This implements a simpler approach to detecting older version of IE

Merges r123630-wpcom.