Sites API: Fix Jetpack wp version for /me/sites Use the new jetpack_wp_version option that is being synced since Jetpack 3.7. If that is not available sync use the wp_version that was synced in jetpack_updates option in Jetpack 3.4 - 3.6.
The reason why we are fixing this now is that we need more accurate data in /me/sites endpoint
to be able to display in the plugin browser where we show the user the compatibility with the plugins and the wp version that the Jetpack site is running.

Merges r123746-wpcom.