Related Posts: resolve divergence from Jetpack
This commit adds filter documentataion, adds one filter for the rel attribute as well
as adds the related posts file to the build script sync tool file list. Note that the
added html_entities_decode call on the excerpt is not merged here because it needs additional
testing.

props @hypertextranch for the changes and the review

This merges the following commits:

    https://github.com/Automattic/jetpack/commit/ef290d5
    https://github.com/Automattic/jetpack/commit/903d473
    https://github.com/Automattic/jetpack/commit/a8a0845
    https://github.com/Automattic/jetpack/commit/4551d74
    https://github.com/Automattic/jetpack/commit/f1dd46b
    https://github.com/Automattic/jetpack/commit/322e13b
    https://github.com/Automattic/jetpack/commit/deca991
    https://github.com/Automattic/jetpack/commit/6602a2d
    https://github.com/Automattic/jetpack/commit/44169f2

Merges r123794-wpcom.