3rd Party: Be exact on file location.
In odd cases, if a specific file calls wp-load.php directly, WP will call wp-settings.php, which loads active plugins. jetpack.php calls the 3rd-party.php which attempt to loads these files. If there is a file in the same directory that matches the names of our files, it will attempt to load that instead of Jetpack's causing gnashing and general sadness.