Contact Form: Add check for `$_POST['sub_menu']` return.
Previously to 4.2, WordPress always assigned a `class="current"` to one of the list table navigation links (`All`, `Spam`, `Trash`). See https://core.trac.wordpress.org/changeset/31828

Now, WP does not assign the class when filtering but we're assuming it is always present. This results in a `E_NOTICE: jetpack/modules/contact-form/admin.php:616 - Undefined index: sub_menu` when performing an AJAX action (spam/not-spam/etc) a feedback when not on the first page of one of the preset views.

Related: The current classes are not added onto the default views if paged forward. Core ticket: https://core.trac.wordpress.org/ticket/33923