Check for absence of `id` when building the registration link
Sometimes on disconnect, the `id` option was removed from the db while the blog_token was left behind.  Since the blog token was still there, we assumed there was an ID and so would get an error when trying to build the registration link based on the ID.  This caused errors, and left users not able to connect.