REST API: Add additional check for date parameter, so future posts dont get scheduled now
A fix so the API doesn't force override a future date with the current time
which negates the scheduling of a post, forcing it to be published NOW.

The checked added in r123566-wpcom only looked for 'date_gmt' field to be empty
and overwrote the date if it was, Postbot was using the 'date' field which from
API documentation looks to be the valid field.

Merges r124419-wpcom.