Check if post date already set, since switching from draft to publish does not require passing in a date
Back out previous change r124419-wpcom and replace with checking date from post object
If the date is in the past then it will be set to now, which is the intent of the first
change that introduced the bug. One caveat, you won't be able to publish a draft post
with a back date (using API). It will have to be published first and then back dated.

References:
r123566-wpcom
r124419-wpcom
https://github.com/Automattic/calypso-pre-oss/issues/9707

Merges r124655-wpcom.