Use `get_post_galleries_images()` in `Jetpack_PostImages::from_gallery()`
Use get_post_galleries_images() instead of a custom functionality.

`get_post_galleries_images()` returns a multidimentional array of image
URLs for all the galleries it finds in a post.

Could have used `get_post_galleries()`, decided against it since it
returns gallery specific data which is not used in this case.

The output is exactly the same as before, except now we're using the
built in WP function get the images.

Fixes #1468