Add heading_level option to portfolio shortcode
The portfolio shortcode currently produces HTML with a 2nd rank heading. For many themes this is not ideal as often the site title is an h1, the page title is an h2, so the logical heading is a h3. However, as this is very theme and use dependant, a change to h3 is not ideal either. To me, it makes sense to add an option to allow the user to control the heading rank should they wish (values restricted to 1--6), but it defaults to 2 so there is minimal impact to existing users.
Can be seen working at http://www.nicholsonweb.co.uk/projects/ to change the heading level to 3.