Add missing parameter to the_title filter call
In core, the filter call for `the_title` provides two parameters: the post title and [the post ID](https://github.com/WordPress/WordPress/blob/master/wp-includes/post-template.php#L156). The filter call in this file was not providing the post ID.

Omitting the ID in this instance of the filter causes a **Warning: Missing argument 2** error for functions that use the second parameter.