REST API: adds revision attribute to /sites/$site/posts/$post endpoint
This will be used in calypso to show inside the editor revisions that have been saved for a given Post and to link to the latest one. This attribute is only added when the Post is requested with ?conext=edit, so as not to slow down our many other posts-list queries.

This was started by @matiasventura in phabricator here https://[private link] and was continued by @rralian in phabricator here https://[private link] (because I wasn't sure how to take over the original review). New tests have been added and confirmed to work correctly. The suite of Post tests have also been run, with a couple of tests failing, however those tests have been consistently failing for a while... nothing new.

Merges r126050-wpcom.