JSON API: add check for `login` scheme
Quoted from @koke - r34213-core changed the action attribute of the login form to use wp_login_url(), which uses the login scheme instead of login_post. This is stripping all the $_GET parameters from the action URL, and the login fails.

This adds a check for both the login and login_post scheme for back-compat