REST API: Media: Support `alt`, `artist`, and `album` attributes in upload
Add support for the alt, artist, and album attributes to be specified in a new upload attempt via the attrs request field. It is claimed in the documentation that these attributes are supported when in fact they do not currently assign the value.

See: https://[private link].1/post/sites/%24site/media/new/
See: D517

Merges r127301-wpcom.