Photon: Thumbnail improvements and enhancements
* Use correct dimensions for older images.
Adds `image_constrain_size_for_editor` to ensure we are not outputting images larger than expected.
* Determine dimensions for uncreated file sizes
For an image size that was not declared when an image was originally created, WordPress does not have any size information for that attachment. Photon now calculates what the dimensions should be.

Photon will now use the correct height/width for all image thumbnails—hard or soft cropped, defined image size at the time of upload or not, as well as custom size arrays passed to get_post_thumbnail(), etc.