JITM: follow AUTOMATIC_UPDATER_DISABLED and auto_update_plugin filter output for current plugin when displaying JITM
If AUTOMATIC_UPDATER_DISABLED constant is set to false, don't display JITM in Updates or Plugins screens.
When activating a plugin:
- if auto_update_plugin filter for such plugin is false, don't display JITM
- if auto updates are already enabled for that plugin in WordPress.com, don't display JITM
- if plugin is not one that can be auto update (it's not in WordPress.org repo), don't display JITM

When one or more plugins are activated, an option 'jetpack_previously_activated' will be saved containing all plugins that were activated before this activation. The option is later compared with 'active_plugins' option. The difference will be those that were just activated. The 'previously_activated' option is then deleted. The plugins that were activated are then checked to find out if they are in WordPress.org repo, if they have WordPress.com auto updates already enabled of if they have an 'auto_update_plugin' filter set upon them.