[Contact Info Widget] [improvement] Removed `extract()` usage. Improved code formatting and style. Added missing PHPDoc.
`extract()` is considered bad practice in the PHP Coding Standards of WordPress. Its usage has been removed from the widget and replaced with proper `$args` usage.

Some of the conditionals didn't use yoda style comparison. Now they do.

A notice was generated when a new instance of the widget was saved, because `$old_instance` was an empty array, as new widget instances don't have previous configuration.

The code formatting has been improved. HTML code in the admin configuration section was a bit broken and it was fixed and formatted. Missing PHPDoc was added to some of the methods.