[Contact Form] Properly check if `$extra_fields` is an array in `get_compiled_form()`
There was no validation of `$extra_fields` in the method and in some edge cases, when the field came back as not an array, an E_NOTICE was generated. The check makes sure that the code that handles extra fields is only executed when the supplied data is an array.