[Display Posts Widget] [improvement] Widget update cron can now be disabled if it is not needed.
Previously if the widget was activated once, there was no way to stop the update cron and it would have been running in the background forever. Functionality has been added to allow cron disabling when it is not needed.

A check is run on every shutdown to see if the cron should be running or not and the appropriate action is taken.