REST API Options endpoint
Allowing for string -> object fallback when we `cast_and_filter_item`
On GET requests for an option_name, and the option_value is an array,
it will always be cast as an object. On POST requests to write an
option, you can now pass a JSON string, and include `is_array` parameter
so that the jetpack site will convert the JSON to a php array.

Merges r129590-wpcom.