[Display Posts Widget] [improvement] Widget update cron is now activated and deactivated in more consistent way.
Code was run on every shutdown of WordPress to check if the cron should have been running or not. This was a great overhead in cases where the widget wasn't needed to be shown.

Now the cron is activated on the proper activation points - When Jetpack gets updated, when Extra Sidebar Widgets module gets activated and when Jetpack gets activated (if it was not activated before).

Deactivation remains what it was before - when Extra Sidebar Widgets module gets deactivated and when Jetpack gets deactivated.