Add base64 for crossdomain workaround
This PR adds base64 encoding to the WOFF font, as modern IE and Firefox both block fonts loading from a separate domain, which will often happen due to CDNs and whatnot. The best way to fix it is to add headers to the domains to allow this, but users of this in a plugin will probably not know to do that. Base64 encoding sidesteps this entirely, as it includes the font in the stylesheet, so the font will always be loaded on the same domain as the stylesheet.

Also rename the font name from "social-logo" to "social-logos", plural.