[Display Posts Widget [fix] Don't show empty URL error for new instances of the widget.
Empty URL error was shown for new instances, which were not yet saved. This means that the error was visible from the moment the widget was dragged from the Available Widgets area to the Widget Area.

Now the error checks if the widget has been already saved, before showing the error.