Jetpack Signature: Check both 80 and 443
Previously, Jetpack used is_ssl() to determine if the port used as part of the signature verification process would be either 80 or 443 (or the corresponding constant values).

With the rise of CloudFlare SSL, Let's Encrypt, etc and the way that HTTPS sites operate when they terminate SSL on the edge (reverse proxy, load balancer, etc), there is an increase of situations where a site is SSL but the server port is 80.

This situation currently always fails and requires user intervention: either overriding the $_SERVER global or setting a PHP constant.

This change will not pass the port number if it is either 80 or 443 while still allowing the constants to be set, to retain BC.