Preventing a db UPDATE query on every page load if it's not needed
Previously the code was blindly issuing an UPDATE against the db
(and thrashing the object cache) on each and every page view for
the main site in a WPMU installation.

This is a variable that is really only ever set once, it makes
absolutely no sense to write to the database on every page view.

Now testing for current value of jetpack_protect_active before
setting it to 1. The call to `get_site_option()` shouldn't add any
overhead because site options are always cached (if I'm not
mistaken).