Update Medium to omit Style tag and Grunt to remove IDs
This commit changes the Gruntfile to clean up IDs, which are useless in a minimized SVG file. In addition to this I noticed a style tag inside the Medium icon source file, which was there because I saved the file wrong from Illustrator. Both of those are fixed.

Incidentally this changes the icon font again since the icon file changed. Visually it won't make a difference.