Customizer: Fix widget visibility button display.
Previously, the widget visibility button did not display for existing widgets in the Customizer; it only appeared on new-added widgets.

The code that moved the button out of widget content and into the widget controls only ran over the prototype widgets in the DOM, and not over the existing widgets loaded in the configuration.

This patch adds the function moveWidgetVisibilityButton and runs it on both load and on the widget-added event. Any widgets that still have the button in their content div will have it moved to the correct control div.

props jmdodd