Tracks: add optional props to the click events
This adds an additional data prop to the elements that if added will be tacked on to the event as a prop.  Else, whatever is put in the data-jptracks-name will be sent over as the name.

Also this adds proper json_success function. This also will not redirect to whatever url it was being linked to before we preventDefault