REST API/Performance: Improve the speed we calculate the post_count and also return just a count of published posts.
post_count calculation is the slowest, the hand crafted query doesn't directly hit and index (the index is on type,status,date) and so can sometimes be slow.
Have seen upper times of over 4 seconds and the 5minute median is 3.3ms and is the slowest part of the response building for a site.

Merges r132149-wpcom.