Infinite Scroll: Hide infinite-scroll class if the option is disabled
The Jetpack support page says that the infinite-scroll class should be used in a theme to hide the navigation links.
However, even when disabled in the Reading page, the class is still visible and the CSS is applied just as if the scroll is enabled.

This commit adds an option check before filtering the body_class classes.

Merges https://github.com/Automattic/jetpack/pull/1208
Props mpeshev

https://[private link]

Merges r132546-wpcom.