Introduce basic JS utilities to work with widgets in Customizer.
They're hooked to the wp object:

wp.customizerHasPartialWidgetRefresh, checks that the Customizer supports partial widget refresh.
wp.isJetpackWidgetPlaced, verifies that the widget placed is the widget that dispatched the event.

To use them and avoid breaking the front end when Customizer is not initialized, they must be used like:

if ( wp && wp.customize && wp.customizerHasPartialWidgetRefresh() ) {
	// do something
}