Easy Markdown: saving post_content_filtered even for autosaves
There is an issue with autosaves as they can't be properly restored as the post_content_filtered is not being populated for them.

If someone attempts to restore an autosave, they would get an empty texteditor, as the content is being read from post_content_filtered.

Populating and storing the post_content_filtered even for autosaves addresses this issue.

/z 50817

Merges r134728-wpcom.