SSO: Fixes profile button styles after #3668
Before #3668, the styles for the button were added in line. So, when we factored out the styles into a separate CSS file, we lost the styles on the profile page. This adds those styles back.

Now that we are using a different style for the login form, it is OK to keep these styles in separate CSS files.