Properly address slashes in custom CSS.
In c2811e253175d8c371739c3e56e7e0d043e471e3, we (I) removed a call to `stripslashes()` that seemed to be interfering with icon fonts.  This caused problems, because other modifications to the CSS expected it to have all slashes removed -- $_POST['safecss'] will be slashed, as WordPress guarantees that, so if we don't run `stripslashes()`, then the CSS will be passed to CSSTidy with slashes in it, causing all sorts of mayhem.

However, if we just run `stripslashes()` and do nothing else, then slashes that are supposed to be in the CSS (like for icon fonts, or quotes inside of quotes) will be stripped by various core WP functions that are expecting slashed data. We pass the CSS to three core functions; `wp_insert_post()` and `wp_update_post()` expect slashed data, but `_wp_put_post_revision()` does not, so we have to account for that. (I am basing whether each expects slashed data on how they modify data that I am passing in, not on any documented slashing requirement.)