Implement General Settings > Miscelaneous Settings
New endpoint /settings returns all settings that are not those of a module.
New endpoint /setting/update allows to update a specific setting.
New component SettingToggle. Uses this.props.children to insert the text label in the component like:
    <SettingToggle>label</SettingToggle> => <FormToggle>{ this.props.children }</FormToggle>
since that's the approach for FormToggle definition in dops components.
Adds a new item to Initial_State JS var: settingNames, which is an array of conflictive option names that are not fixed. Right now only the holiday snow setting is in it, since the option name is run through a filter and can only be get through the jetpack_holiday_snow_option_name() function.