Create new single use upload tokens to enable VideoPress uploads from Jetpack
Summary:
Main summary is located here: https://[private link]

The short version is:

This adds in the idea of a single use authentication token that can be requested via a X_JETPACK token from the REST API. It uses Memcached to store temporary tokens that allow a user to perform a "site" action once. As soon as the token is used for authentication, it is destroyed. The idea is to enable a way for Jetpack sites to request a temporary token that can then be used via Javascript to upload a video to the REST api.

Test Plan:
Main testing for this has been done via 2 methods.

The first was to create a small PHP application that acts as a console version of Jetpack, which would call to the REST API using the public/private Jetpack tokens to get an upload token, then use that token to upload a video. This was used to verify that the upload tokens can only be retrieved for 1) A valid X_JETPACK token 2) The Jetpack blog_id the X_JETPACK token is tied to 3) Any VideoPress blogs that the X_JETPACK token is tied to via its users. Once the token was verified, it was then used to upload a video.

The second was to use the new branch on Jetpack (https://github.com/Automattic/jetpack/tree/feature/videopress-uploader) to upload a video from the media uploader, to the REST API.

Both tests worked.

Conditions tested so far:

1) The upload token only gives access to the Media Upload endpoint
2) The upload token is usable only once
3) The upload token only works for the blog_id that it was whitelisted for
4) An upload token can only be retrieved for a blog that the authenticated user (or Jetpack token) has access to.

Reviewers: #api, georgestephanis

Differential Revision: https://[private link]

Merges r135573-wpcom.