Endpoint: Dismiss Jetpack Notice
This new endpoint will handle dismissing of any Jetpack notice shown in the js app.  It introduces a new option `jetapck_dismissed_notices` that will be stored as an object and passed initally to the client via window.Initial_State.dismissedNotices.

On dismissal of a notice, it will update the option.

New filter added `jetpack_dismissed_notices` that can be used to control which are dismissed.